+++ backend/windmill-worker/src/common.rs
@@ -43,7 +43,6 @@ use windmill_common::{variables, DB};
 use tokio::{io::AsyncWriteExt, time::Instant};
 
 use crate::agent_workers::UPDATE_PING_URL;
-use crate::utils::get_root_job_id;
 use crate::{JOB_DEFAULT_TIMEOUT, MAX_RESULT_SIZE, MAX_TIMEOUT_DURATION, PATH_ENV};
 use windmill_common::client::AuthedClient;
 
@@ -1051,6 +1050,14 @@ pub fn build_http_client(timeout_duration: std::time::Duration) -> error::Result
         .connect_timeout(std::time::Duration::from_secs(10))
         .build()
         .map_err(|e| Error::internal_err(format!("Error building http client: {e:#}")))
+}
+
+pub fn get_root_job_id(job: &MiniPulledJob) -> uuid::Uuid {
+    // fallback to flow_innermost_root_job and parent_job as root_job is not set if equal to innermost root job or parent job
+    job.root_job
+        .or(job.flow_innermost_root_job)
+        .or(job.parent_job)
+        .unwrap_or(job.id)
 }
 
 #[derive(Clone)]
